dht, dht/dhtutil: add util package for helper functions
