+++ dht/dhtutil/resolver_test.go
@@ -1,0 +1,1 @@
+package dhtutil_test
